+++ src/client/java/com/orelizards/client/OreLizardModel.java
@@ -4,9 +4,9 @@ import com.orelizards.OreLizardsMod;
 import com.orelizards.entity.OreLizardEntity;
 import net.minecraft.resources.Identifier;
 import org.jetbrains.annotations.Nullable;
-import software.bernie.geckolib.constant.dataticket.DataTicket;
-import software.bernie.geckolib.model.GeoModel;
-import software.bernie.geckolib.renderer.base.GeoRenderState;
+import com.geckolib.constant.dataticket.DataTicket;
+import com.geckolib.model.GeoModel;
+import com.geckolib.renderer.base.GeoRenderState;
 
 public class OreLizardModel extends GeoModel<OreLizardEntity> {
 	private static final Identifier MODEL =